+++ src/user-preferences/user-preferences.service.ts
@@ -1,10 +1,16 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import {
+  ForbiddenException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { CreateUserPreferenceDto } from './dto/create-user-preference.dto';
-import { UpdateUserPreferenceDto } from './dto/update-user-preference.dto';
 import { UserPreference } from './entities/user-preference.entity';
 import { LoggerService } from '../common/logger/logger.service';
+import { IUpdateUserPreference } from './types';
+import { RoleName } from '../roles/entities/role.entity';
+import { User } from '../users/entities/user.entity';
 
 @Injectable()
 export class UserPreferencesService {
@@ -71,14 +77,36 @@ export class UserPreferencesService {
     return preference;
   }
 
-  async update(
-    userId: string,
-    updateUserPreferenceDto: UpdateUserPreferenceDto,
-  ): Promise<UserPreference> {
+  async update({
+    currentUser,
+    userId,
+    updateUserPreferenceDto,
+  }: IUpdateUserPreference): Promise<UserPreference> {
     this.logger.log(`Updating preferences for user ${userId}`, {
       context: UserPreferencesService.name,
       meta: { updateUserPreferenceDto, userId },
     });
+
+    const isAdmin = currentUser.roles?.some(
+      (role) => role.name === RoleName.ADMIN,
+    );
+
+    if (!isAdmin && currentUser.id !== userId) {
+      this.logger.warn(
+        `User ${currentUser.id} tried to update preferences ${userId} without permission`,
+        {
+          context: UserPreferencesService.name,
+          meta: {
+            currentUserId: currentUser.id,
+            targetUserId: userId,
+          },
+        },
+      );
+
+      throw new ForbiddenException(
+        "You are not allowed to update this user's preference",
+      );
+    }
 
     const preference = await this.findOne(userId);
 
@@ -97,11 +125,25 @@ export class UserPreferencesService {
     return saved;
   }
 
-  async remove(userId: string): Promise<{ message: string }> {
+  async remove(
+    currentUser: User,
+    userId: string,
+  ): Promise<{ message: string }> {
     this.logger.log(`Removing preferences for user ${userId}`, {
       context: UserPreferencesService.name,
       meta: { userId },
     });
+
+    const isAdmin = currentUser.roles?.some(
+      (role) => role.name === RoleName.ADMIN,
+    );
+
+    if (!isAdmin && currentUser.id !== userId) {
+      throw new ForbiddenException(
+        'You are not allowed to delete this preferences',
+      );
+    }
+
     const result = await this.userPreferenceRepository.delete(userId);
 
     if (result.affected === 0) {
